Refactor Quiz to take a map of tenses and their corresponding allowed subjects
Previously, Quiz took a separate `tenses` and `subjects` arg and would
randomly pick from each to generate the subject–tense combination in
which the verb would be conjugated. Now, Quiz takes a map wherein each
key is a tense and each value is an array of allowed subjects for that
tense. This is in awknowledgment of the fact that some tenses can’t be
conjugated in some subjects (e.g. imperativo in io).

Finally, disable/enableTense methods were added. Previously, a tense
would be disabled simply by removing it from the tenses property. The
new approach is cleaner for internally managing the new tense map, and
it offers better encapsulation/forwards compability anyway.