Refactor Quiz
Make clean answer a “static method” on the Quiz constructor, since it
was operating algorithmically (i.e. not relying on `this` at all)
anyway.

Merge gradeAnswer into recordAnswer, since grading is part of
recording, and this makes the interface simpler for clients. Also,
create a getPointsForAnswer method, which is the only kind of “grading”
that can really be called such outside of recording.

Set recordAnswer to only unset the currentVerb if the quiz is done;
otherwise, the currentVerb remains as is until nextPrompt is called.