New build process
Precompile the mustache templates and the coffeescript, before feeding
all the resulting js to the requirejs optimizer, which builds a final
script. Also minify the CSS. Then clean out the build folder of the
temporary scripts made along the way and update index.html, which is
copied into build too, to point to the new CSS and the new requirejs
files.